update: 更新文件 isshowableasciionly_fuzzer.cpp

+++ test/fuzztest/isshowableasciionly_fuzzer/isshowableasciionly_fuzzer.cpp
@@ -34,10 +34,10 @@ void DoSomethingInterestingWithMyAPI(const uint8_t *data, size_t size)
     auto simUtils = std::make_shared<SIMUtils>();
     int32_t byteslen = static_cast<int32_t>(*data);
     char argument = static_cast<char>(*data);
-    int32_t offset = 0;
+    size_t offset = 0;
     std::string str(reinterpret_cast<const char *>(data), size);
     offset += sizeof(int32_t);
-    offset = (offset > static_cast<int32_t>size) ? static_cast<int32_t>size : offset;
+    offset = (offset > size) ? size : offset;
     std::string parameter(reinterpret_cast<const char *>(data + offset), size - offset);
     simUtils->HexStringConvertToBytes(str, byteslen);
     simUtils->IsShowableAsciiOnly(str);
